% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_note.R
\name{add_note}
\alias{add_note}
\title{Quickly record a note for future reference}
\usage{
add_note(note, name = "general", ext = "Rmd")
}
\arguments{
\item{note}{The actual note in quotes}

\item{name}{The name of the notes file that you want to create or add to.
Default is "general"}

\item{ext}{The file extension for the notes file. Default is "Rmd"}
}
\value{
Creates or adds to a file named \code{name} with contents
\code{note}.
}
\description{
Quickly record a note for future reference
}
\examples{
add_note("Remember to tweak the tuning grid for the random forest model",
name = "random-forest")

}
\seealso{
\code{\link{add_notes}} for capturing many notes from an R script.
}
