% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_notes.R
\name{add_notes}
\alias{add_notes}
\title{Quickly record many notes from current R script (not working)}
\usage{
add_notes()
}
\value{
Notes in the specified location.
}
\description{
\code{add_notes} reads the current active script and looks for special
comment markers #!# and records everything between those markers.
}
\examples{
add_notes()
}
