% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_values.R
\name{count_values}
\alias{count_values}
\title{Count values in multiple columns}
\usage{
count_values(data, ...)
}
\arguments{
\item{data}{A data.frame or tibble}

\item{...}{Columns that you want to count the values of}
}
\value{

}
\description{
Count values in multiple columns
}
\examples{
count_values(mtcars, carb, vs, am)
}
