% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_started.R
\name{get_started}
\alias{get_started}
\title{Create frequently used directories}
\usage{
get_started()
}
\value{

}
\description{
Use this function when starting a new project to help keep code and files
organized. Following a project-oriented workflow, the working directory
should be set as the project directory.
get_started creates the following files:
}
\details{
\itemize{
\item data-raw
\item data-derived
\item eda
\item scripts
\item models
\item helpers
}
}
\examples{
get_started()
}
