% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_theme_toggle.R
\name{get_theme_toggle}
\alias{get_theme_toggle}
\title{Toggle between light and dark mode in RStudio}
\usage{
get_theme_toggle()
}
\value{
Updated global .Rprofile with code to help you toggle between
light and dark mode.
}
\description{
\code{get_theme_toggle} writes some code to your global .Rprofile file so
that you can more easily set up the ability to toggle between light and
dark themes in RStudio.
}
\examples{
get_theme_toggle()
}
