% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_directories.R
\name{ls_directories}
\alias{ls_directories}
\alias{ls_models}
\alias{ls_scripts}
\alias{ls_eda}
\alias{ls_data_derived}
\alias{ls_data_raw}
\alias{ls_helpers}
\title{Simple functions to list the contents of a specified directory}
\usage{
ls_models()

ls_scripts()

ls_eda()

ls_data_derived()

ls_data_raw()

ls_helpers()
}
\value{
a character vector of files
}
\description{
Create a new or open an existing script in one of three directories:
models, scripts, or eda.
}
\section{Functions}{
\itemize{
\item \code{ls_models}: List all of the scripts in the models directory

\item \code{ls_scripts}: List all of the scripts in the scripts directory

\item \code{ls_eda}: List all of the scripts in the eda directory

\item \code{ls_data_derived}: List all of the scripts in the data-derived
directory

\item \code{ls_data_raw}: List all of the scripts in the data-raw directory

\item \code{ls_helpers}: List all of the scripts in the helpers directory
}}

\examples{
ls_models()
ls_scripts()
ls_eda
}
