% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_eda.R
\name{new_eda}
\alias{new_eda}
\title{Create a new script in eda}
\usage{
new_eda(script_name, author = "Avery Robbins")
}
\arguments{
\item{script_name}{a character vector}

\item{author}{a character vector}
}
\value{
a new script in eda
}
\description{
Create a new script in eda
}
\examples{
new_eda("01_initial_look")
}
