% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_model.R
\name{new_model}
\alias{new_model}
\title{Create a new script in models}
\usage{
new_model(script_name, author = "Avery Robbins")
}
\arguments{
\item{script_name}{a character vector}

\item{author}{a character vector}
}
\value{
a new script in models
}
\description{
Create a new script in models
}
\examples{
new_model("01_random_forest")
}
