% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline_workflow.R
\name{outline_workflow}
\alias{outline_workflow}
\title{Provide a useful outline of a typical tidymodels workflow}
\usage{
outline_workflow()
}
\value{
a character vector
}
\description{
Provide a useful outline of a typical tidymodels workflow
}
