% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_all.R
\name{print_all}
\alias{print_all}
\title{Print every row of data}
\usage{
print_all(data)
}
\arguments{
\item{data}{Tibble, data.frame, etc. that you want to print.}
}
\value{
A dataset with all of the rows printed to screen.
}
\description{
Quickly print every row of your data.
}
\examples{
`\%>\%` <- magrittr::`\%>\%`
dplyr::starwars \%>\% dplyr::count(name) \%>\% print_all()
}
