% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{prop}
\alias{prop}
\title{Quickly calculate the proportion of values in a column}
\usage{
prop(data, column, na.rm = TRUE)
}
\arguments{
\item{data}{a tibble or data.frame}

\item{column}{a column to be counted}

\item{na.rm}{a logical to remove NAs or not}
}
\value{
a tibble
}
\description{
Quickly calculate the proportion of values in a column
}
\examples{
prop(mtcars, am)
}
