% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remember_package.R
\name{remember_package}
\alias{remember_package}
\title{Keep a record of packages}
\usage{
remember_package(package_name)
}
\arguments{
\item{package_name}{The package that you want to remember surrounded
in quotes.}
}
\value{

}
\description{
Keep a record of packages
}
\examples{
remember_package("ggthemr")
}
