% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_helpers.R
\name{source_helpers}
\alias{source_helpers}
\title{Source helper functions}
\usage{
source_helpers()
}
\value{

}
\description{
Quickly source all of the files in a directory called 'helpers'. The
'helpers' directory should be a top level directory inside the project
directory.
}
\examples{
source_helpers()
}
