% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_files.R
\name{use_files}
\alias{use_files}
\alias{use_model}
\alias{use_script}
\alias{use_eda}
\title{Functions to create or open specific files}
\usage{
use_model(script_name, author = "Avery Robbins")

use_script(script_name, author = "Avery Robbins")

use_eda(script_name, author = "Avery Robbins")
}
\arguments{
\item{script_name}{a character vector}

\item{author}{a character vector}
}
\value{
a .R file in a specified directory
}
\description{
Create a new or open an existing script in one of three directories:
models, scripts, or eda.
}
\section{Functions}{
\itemize{
\item \code{use_model}: Create a script for modeling

\item \code{use_script}: Create a new or open an existing script for
data processing

\item \code{use_eda}: Create a new or open an existing script for eda
}}

\examples{
use_model("01_random-forest")
use_script("monthly-project")
use_eda("explore-canvas")
}
