% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_script.R
\name{use_script}
\alias{use_script}
\title{Create a new script in scripts}
\usage{
use_script(script_name, author = "Avery Robbins")
}
\arguments{
\item{script_name}{a character vector}

\item{author}{a character vector}
}
\value{
a new script in scripts
}
\description{
Create a new script in scripts
}
\examples{
use_script("01_student-total")
}
