% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_started.R
\name{get_started}
\alias{get_started}
\title{Create frequently used directories}
\usage{
get_started()
}
\value{

}
\description{
Use this function when starting a new project to help keep code and files
organized. Following a project-oriented workflow, the working directory
should be set as the project directory.
}
\details{
get_started creates the following directories: data-raw, data-derived, eda,
scripts, models, & helpers.

See \code{\link{get_started2}} for greater customization options.
}
\examples{
get_started()
}
