% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_started2.R
\name{get_started2}
\alias{get_started2}
\title{Create custom directories easily}
\usage{
get_started2(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Create whatever directories you want. Type in the directory names in quotes
separated by commas.
}
\examples{
get_started2("data", "reports", "analysis")
}
